requirements.txt
dj-database-url==0.5.0
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
pytz==2018.4
